+++ apps/eew/scfinder/main.cpp
@@ -215,10 +215,10 @@ class App : public Client::StreamApplication {
 			_sentMessagesTotal = 0;
 			_testMode = false;
 			_playbackMode = false;
-			_defaultPGWindowLength = 30
 
 			_bufferLength = Core::TimeSpan(120,0);
-			_PGWindowLength = Core::TimeSpan(_defaultPGWindowLength,0);
+			_bufDefaultLen = Core::TimeSpan(60,0);
+			_bufVarLen = Core::TimeSpan(60,0);
 
 			// Default Finder call interval is 1s
 			_finderProcessCallInterval.set(1);
@@ -315,6 +315,12 @@ class App : public Client::StreamApplication {
 
 			try {
 				_bufferLength = configGetDouble("finder.envelopeBufferSize");
+			}
+			catch ( ... ) {}
+
+			try {
+				_bufDefaultLen = configGetDouble("finder.defaultFinDerEnvelopeLength");
+				_bufVarLen = configGetDouble("finder.defaultFinDerEnvelopeLength");
 			}
 			catch ( ... ) {}
 
@@ -563,7 +569,8 @@ class App : public Client::StreamApplication {
 			}
 
 			// The minimum time for a valid amplitude
-			Core::Time minAmplTime = _referenceTime - _bufferLength;
+			//Core::Time minAmplTime = _referenceTime - _bufferLength;
+			Core::Time minAmplTime = _referenceTime - _bufVarLen;
 
 			#if defined(LOG_AMPS)
 			std::cout << "+ " << id << "." << proc->waveformID().channelCode() << "   " << _referenceTime.iso() << "   " << minAmplTime.iso() << "   " << timestamp.iso() << "   " << value << std::endl;
@@ -643,13 +650,13 @@ class App : public Client::StreamApplication {
 			for ( it = _locationLookup.begin(); it != _locationLookup.end(); ++it ) {
 				if ( !it->second->maxPGA.timestamp.valid() ) continue;
 
-				if ( now - it->second->maxPGA.timestamp > _PGWindowLength ) {
+				if ( now - it->second->maxPGA.timestamp > _bufVarLen ) {
 					SEISCOMP_DEBUG("PGA %s.%s.%s.%s older than %s s has been ignored", 
 								   it->second->meta->station()->code(),
 								   it->second->meta->station()->network()->code(),
 								   it->second->maxPGA.channel.c_str(),
 								   it->second->meta->code().empty()?"--":it->second->meta->code().c_str(),
-								   _PGWindowLength.c_str());
+								   _bufVarLen.c_str());
 					continue
 				}
 				if ( it->second->maxPGA.clipped ) {}
@@ -732,6 +739,7 @@ class App : public Client::StreamApplication {
 
 			#ifdef USE_FINDER
 			Finder_List::iterator fit;
+			double maxRupLen = 0.;
 			for ( fit = _finderList.begin(); fit != _finderList.end(); /* incrementing below */) {
 				// some method for getting the timestamp associated with the data
 				// event_continue == false when we want to stop processing
@@ -740,6 +748,9 @@ class App : public Client::StreamApplication {
 				}
 				catch ( FiniteFault::Error &e ) {
 					SEISCOMP_ERROR("Exception from FinDer::process: %s", e.what());
+				}
+				if ((*fit)->get_rupture_length() > maxRupLen) {
+					maxRupLen = (*fit)->get_rupture_length();
 				}
 
 				if ( (*fit)->get_finder_flags().get_message() &&
@@ -754,6 +765,20 @@ class App : public Client::StreamApplication {
 				}
 				else
 					++fit;
+			}
+			if (_finderList.size() == 0) {
+				if (_bufVarLen != _bufDefaultLen) {
+					SEISCOMP_DEBUG("Resetting data window to %ld", _bufDefaultLen.seconds());
+				}
+				_bufVarLen = _bufDefaultLen;
+			} else {
+				double rup2time = 1.5;
+				if (maxRupLen > _bufVarLen * rup2time) {
+					double tmp = maxRupLen / rup2time;
+					_bufVarLen = min((long)tmp, _bufferLength.seconds());
+					SEISCOMP_DEBUG("Increasing data window to %ld because of active FinDer event rupture length %.1f", 
+					  _bufVarLen.seconds(), maxRupLen);
+				}
 			}
 			#else
 			std::cerr << "ProcessData" << std::endl;
@@ -1042,9 +1067,8 @@ class App : public Client::StreamApplication {
 		std::string                    _finderConfig;
 
 		Core::TimeSpan                 _bufferLength;
-
-		int 						   _defaultPGWindowLength
-		Core::TimeSpan                 _PGWindowLength;
+		Core::TimeSpan                 _bufDefaultLen;
+		Core::TimeSpan                 _bufVarLen;
 
 		Processing::EEWAmps::Processor _eewProc;
 		CreationInfo                   _creationInfo;